function display_cover
The function display_cover selects a random image from https://picsum.photos/ and will help us superimpose two strings over the image. The parameter top is the string we would like to superimpose on the top of an image. The parameter bottom is the string we would like to display on the bottom of the image. The function does not return the image but returns an object of type Image from the Pillow library; the object represents a PIL image.